Remove redundant YAML modification - LD_LIBRARY_PATH already set in shell
The LD_LIBRARY_PATH environment variable is already exported globally
in the shell session (line 270), so we don't need to add it to the
compiler configuration in packages.yaml. This was causing YAML
formatting issues and is unnecessary since the env var is inherited
by all child processes including Spack builds.